generate_students_api/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
faker>=19.6.2
pydantic>=2.3.0
python-multipart>=0.0.6
